Hedgehog.Internal.Config

Documentation

data UseColor

Whether to render output using ANSI colors or not.

Constructors

DisableColor
Disable ANSI colors in report output.

EnableColor
Enable ANSI colors in report output.

resolveColor :: MonadIO m => Maybe UseColor -> m UseColor

data Verbosity

How verbose should the report output be.

Constructors

Quiet
Only display the summary of the test run.

Normal
Display each property as it is running, as well as the summary.

resolveVerbosity :: MonadIO m => Maybe Verbosity -> m Verbosity

newtype WorkerCount

The number of workers to use when running properties in parallel.

Constructors

WorkerCount Int

resolveWorkers :: MonadIO m => Maybe WorkerCount -> m WorkerCount

detectMark :: MonadIO m => m Bool

detectColor :: MonadIO m => m UseColor

detectVerbosity :: MonadIO m => m Verbosity

detectWorkers :: MonadIO m => m WorkerCount