Hedgehog.Internal.Tree

Documentation

type Tree = TreeT Identity

A rose tree.

pattern Tree :: NodeT Identity a -> Tree a

Pattern to ease construction / deconstruction of pure trees.

newtype TreeT m a

An effectful tree, each node in the tree can have an effect before it is produced.

Constructors

TreeT

Fields

runTreeT :: m (NodeT m a)

runTree :: Tree a -> Node a

Extracts the Node from a Tree.

mapTreeT :: (m (NodeT m a) -> m (NodeT m a)) -> TreeT m a -> TreeT m a

Map between TreeT computations.

treeValue :: Tree a -> a

The value at the root of the Tree.

treeChildren :: Tree a -> [Tree a]

The children of the Tree.

type Node = NodeT Identity

A node in a rose tree.

pattern Node :: a -> [Tree a] -> Node a

Pattern to ease construction / deconstruction of pure nodes.

data NodeT m a

A node in an effectful tree, as well as its unevaluated children.

Constructors

NodeT

Fields

nodeValue :: a

The value at this NodeT in the TreeT.

nodeChildren :: [TreeT m a]

The children of this NodeT.

fromNodeT :: Applicative m => NodeT m a -> TreeT m a

Create a TreeT from a NodeT

unfold :: Monad m => (a -> [a]) -> a -> TreeT m a

Create a tree from a value and an unfolding function.

unfoldForest :: Monad m => (a -> [a]) -> a -> [TreeT m a]

Create a forest from a value and an unfolding function.

expand :: Monad m => (a -> [a]) -> TreeT m a -> TreeT m a

Expand a tree using an unfolding function.

prune :: Monad m => Int -> TreeT m a -> TreeT m a

Throw away n levels of a tree's children.

prune 0 will throw away all of a tree's children.

catMaybes :: Tree (Maybe a) -> Maybe (Tree a)

Takes a tree of Maybes and returns a tree of all the Just values.

If the root of the tree is Nothing then Nothing is returned.

filter :: (a -> Bool) -> Tree a -> Maybe (Tree a)

Returns a tree containing only elements that match the predicate.

If the root of the tree does not match the predicate then Nothing is returned.

mapMaybe :: (a -> Maybe b) -> Tree a -> Maybe (Tree b)

filterMaybeT :: (a -> Bool) -> TreeT (MaybeT Identity) a -> TreeT (MaybeT Identity) a

Returns a tree containing only elements that match the predicate.

If the root of the tree does not match the predicate then Nothing is returned.

mapMaybeMaybeT :: (a -> Maybe b) -> TreeT (MaybeT Identity) a -> TreeT (MaybeT Identity) b

filterT :: (Monad m, Alternative m) => (a -> Bool) -> TreeT m a -> TreeT m a

Returns a tree containing only elements that match the predicate.

When an element does not match the predicate its node is replaced with empty.

mapMaybeT :: (Monad m, Alternative m) => (a -> Maybe b) -> TreeT m a -> TreeT m b

depth :: Tree a -> Int

Returns the depth of the deepest leaf node in the tree.

interleave :: forall m a. Monad m => [NodeT m a] -> NodeT m [a]

render :: Tree String -> String

Render a tree of strings.

renderT :: Monad m => TreeT m String -> m String

Render a tree of strings, note that this forces all the delayed effects in the tree.