Hedgehog

Description

This module includes almost everything you need to get started writing property tests with Hedgehog.

It is designed to be used alongside Hedgehog.Gen and Hedgehog.Range, which should be imported qualified. You also need to enable Template Haskell so the Hedgehog test runner can find your properties.

{-# LANGUAGE TemplateHaskell #-}

import           Hedgehog
import qualified Hedgehog.Gen as Gen
import qualified Hedgehog.Range as Range

Once you have your imports set up, you can write a simple property:

prop_reverse :: Property
prop_reverse =
  property $ do
    xs <- forAll $ Gen.list (Range.linear 0 100) Gen.alpha
    reverse (reverse xs) === xs

And add the Template Haskell splice which will discover your properties:

tests :: IO Bool
tests =
  checkParallel $$(discover)

If you prefer to avoid macros, you can specify the group of properties to run manually instead:

{-# LANGUAGE OverloadedStrings #-}

tests :: IO Bool
tests =
  checkParallel $ Group "Test.Example" [
      ("prop_reverse", prop_reverse)
    ]

You can then load the module in GHCi, and run it:

λ tests
━━━ Test.Example ━━━
  ✓ prop_reverse passed 100 tests.

Properties

data Property

A property test, along with some configurable limits like how many times to run the test.

data PropertyT m a

The property monad transformer allows both the generation of test inputs and the assertion of expectations.

data Group

A named collection of property tests.

Constructors

Group

Fields

groupName :: !GroupName

groupProperties :: ![(PropertyName, Property)]

data PropertyName

The name of a property.

Can be constructed using OverloadedStrings:

  "apples" :: PropertyName

data GroupName

The name of a group of properties.

Can be constructed using OverloadedStrings:

  "fruit" :: GroupName

property :: HasCallStack => PropertyT IO () -> Property

Creates a property with the default configuration.

test :: Monad m => TestT m a -> PropertyT m a

Lift a test in to a property.

Because both TestT and PropertyT have MonadTest instances, this function is not often required. It can however be useful for writing functions directly in TestT and thus gaining a MonadTransControl instance at the expense of not being able to generate additional inputs using forAll.

One use case for this is writing tests which use ResourceT:

  property $ do
    n <- forAll $ Gen.int64 Range.linearBounded
    test . runResourceT $ do
      -- test with resource usage here

forAll :: (Monad m, Show a, HasCallStack) => Gen a -> PropertyT m a

Generates a random input for the test by running the provided generator.

forAllWith :: (Monad m, HasCallStack) => (a -> String) -> Gen a -> PropertyT m a

Generates a random input for the test by running the provided generator.

This is a the same as forAll but allows the user to provide a custom rendering function. This is useful for values which don't have a Show instance.

discard :: Monad m => PropertyT m a

Discards the current test entirely.

check :: MonadIO m => Property -> m Bool

Check a property.

recheck :: MonadIO m => Size -> Seed -> Property -> m ()

Check a property using a specific size and seed.

discover :: TExpQ Group

Discover all the properties in a module.

Functions starting with prop_ are assumed to be properties.

checkParallel :: MonadIO m => Group -> m Bool

Check a group of properties in parallel.

Warning: although this check function runs tests faster than checkSequential, it should be noted that it may cause problems with properties that are not self-contained. For example, if you have a group of tests which all use the same database table, you may find that they interfere with each other when being run in parallel.

Using Template Haskell for property discovery:

tests :: IO Bool
tests =
  checkParallel $$(discover)

With manually specified properties:

tests :: IO Bool
tests =
  checkParallel $ Group "Test.Example" [
      ("prop_reverse", prop_reverse)
    ]

checkSequential :: MonadIO m => Group -> m Bool

Check a group of properties sequentially.

Using Template Haskell for property discovery:

tests :: IO Bool
tests =
  checkSequential $$(discover)

With manually specified properties:

tests :: IO Bool
tests =
  checkSequential $ Group "Test.Example" [
      ("prop_reverse", prop_reverse)
    ]

withTests :: TestLimit -> Property -> Property

Set the number of times a property should be executed before it is considered successful.

If you have a test that does not involve any generators and thus does not need to run repeatedly, you can use withTests 1 to define a property that will only be checked once.

data TestLimit

The number of successful tests that need to be run before a property test is considered successful.

Can be constructed using numeric literals:

  200 :: TestLimit

withDiscards :: DiscardLimit -> Property -> Property

Set the number of times a property is allowed to discard before the test runner gives up.

data DiscardLimit

The number of discards to allow before giving up.

Can be constructed using numeric literals:

  10000 :: DiscardLimit

withShrinks :: ShrinkLimit -> Property -> Property

Set the number of times a property is allowed to shrink before the test runner gives up and prints the counterexample.

data ShrinkLimit

The number of shrinks to try before giving up on shrinking.

Can be constructed using numeric literals:

  1000 :: ShrinkLimit

withRetries :: ShrinkRetries -> Property -> Property

Set the number of times a property will be executed for each shrink before the test runner gives up and tries a different shrink. See ShrinkRetries for more information.

data ShrinkRetries

The number of times to re-run a test during shrinking. This is useful if you are testing something which fails non-deterministically and you want to increase the change of getting a good shrink.

If you are doing parallel state machine testing, you should probably set shrink retries to something like 10. This will mean that during shrinking, a parallel test case requires 10 successful runs before it is passes and we try a different shrink.

Can be constructed using numeric literals:

  0 :: ShrinkRetries

Generating Test Data

type Gen = GenT Identity

Generator for random values of a.

data GenT m a

Monad transformer which can generate random values of a.

class Monad m => MonadGen m where

Class of monads which can generate input data for tests.

The functions on this class can, and should, be used without their Gen suffix by importing Hedgehog.Gen qualified.

Methods

liftGen :: Gen a -> m a

See Gen.lift

shrinkGen :: (a -> [a]) -> m a -> m a

See Gen.shrink

pruneGen :: m a -> m a

See Gen.prune

scaleGen :: (Size -> Size) -> m a -> m a

See Gen.scale

freezeGen :: m a -> m (a, m a)

See Gen.freeze

data Range a

A range describes the bounds of a number to generate, which may or may not be dependent on a Size.

The constructor takes an origin between the lower and upper bound, and a function from Size to bounds. As the size goes towards 0, the values go towards the origin.

newtype Size

Tests are parameterized by the size of the randomly-generated data, the meaning of which depends on the particular generator used.

Constructors

Size

Fields

unSize :: Int

data Seed

A splittable random number generator.

Constructors

Seed

Fields

seedValue :: !Word64

seedGamma :: !Word64

must be an odd number

Tests

type Test = TestT Identity

A test monad allows the assertion of expectations.

data TestT m a

A test monad transformer allows the assertion of expectations.

class Monad m => MonadTest m where

Methods

liftTest :: Test a -> m a

annotate :: (MonadTest m, HasCallStack) => String -> m ()

Annotates the source code with a message that might be useful for debugging a test failure.

annotateShow :: (MonadTest m, Show a, HasCallStack) => a -> m ()

Annotates the source code with a value that might be useful for debugging a test failure.

footnote :: MonadTest m => String -> m ()

Logs a message to be displayed as additional information in the footer of the failure report.

footnoteShow :: (MonadTest m, Show a) => a -> m ()

Logs a value to be displayed as additional information in the footer of the failure report.

success :: MonadTest m => m ()

Another name for pure ().

failure :: (MonadTest m, HasCallStack) => m a

Causes a test to fail.

assert :: (MonadTest m, HasCallStack) => Bool -> m ()

Fails the test if the condition provided is False.

(===) :: (MonadTest m, Eq a, Show a, HasCallStack) => a -> a -> m () infix 4

Fails the test if the two arguments provided are not equal.

(/==) :: (MonadTest m, Eq a, Show a, HasCallStack) => a -> a -> m () infix 4

Fails the test if the two arguments provided are equal.

tripping :: (MonadTest m, Applicative f, Show b, Show (f a), Eq (f a), HasCallStack) => a -> (a -> b) -> (b -> f a) -> m ()

Test that a pair of encode / decode functions are compatible.

eval :: (MonadTest m, HasCallStack) => a -> m a

Fails the test if the value throws an exception when evaluated to weak head normal form (WHNF).

evalM :: (MonadTest m, MonadCatch m, HasCallStack) => m a -> m a

Fails the test if the action throws an exception.

The benefit of using this over simply letting the exception bubble up is that the location of the closest evalM will be shown in the output.

evalIO :: (MonadTest m, MonadIO m, HasCallStack) => IO a -> m a

Fails the test if the IO action throws an exception.

The benefit of using this over liftIO is that the location of the exception will be shown in the output.

evalEither :: (MonadTest m, Show x, HasCallStack) => Either x a -> m a

Fails the test if the Either is Left, otherwise returns the value in the Right.

evalExceptT :: (MonadTest m, Show x, HasCallStack) => ExceptT x m a -> m a

Fails the test if the ExceptT is Left, otherwise returns the value in the Right.

State Machine Tests

data Command n m (state :: (* -> *) -> *)

The specification for the expected behaviour of an Action.

Constructors

(HTraversable input, Show (input Symbolic), Typeable output) => Command

Fields

commandGen :: state Symbolic -> Maybe (n (input Symbolic))

A generator which provides random arguments for a command. If the command cannot be executed in the current state, it should return Nothing.

commandExecute :: input Concrete -> m output

Executes a command using the arguments generated by commandGen.

commandCallbacks :: [Callback input output state]

A set of callbacks which provide optional command configuration such as pre-condtions, post-conditions and state updates.

data Callback input output state

Optional command configuration.

Constructors

Require (state Symbolic -> input Symbolic -> Bool)
A pre-condition for a command that must be verified before the command can be executed. This is mainly used during shrinking to ensure that it is still OK to run a command despite the fact that some previously executed commands may have been removed from the sequence.

Update (forall v. Ord1 v => state v -> input v -> Var output v -> state v)
Updates the model state, given the input and output of the command. Note that this function is polymorphic in the type of values. This is because it must work over Symbolic values when we are generating actions, and Concrete values when we are executing them.

Ensure (state Concrete -> state Concrete -> input Concrete -> output -> Test ())
A post-condition for a command that must be verified for the command to be considered a success.

This callback receives the state prior to execution as the first argument, and the state after execution as the second argument.

data Action m (state :: (* -> *) -> *)

An instantiation of a Command which can be executed, and its effect evaluated.

data Sequential m state

A sequence of actions to execute.

Constructors

Sequential

Fields

sequentialActions :: [Action m state]

The sequence of actions.

data Parallel m state

A sequential prefix of actions to execute, with two branches to execute in parallel.

Constructors

Parallel

Fields

parallelPrefix :: [Action m state]

The sequential prefix.

parallelBranch1 :: [Action m state]

The first branch.

parallelBranch2 :: [Action m state]

The second branch.

executeSequential :: (MonadTest m, MonadCatch m, HasCallStack) => (forall v. state v) -> Sequential m state -> m ()

Executes a list of actions sequentially, verifying that all post-conditions are met and no exceptions are thrown.

To generate a sequence of actions to execute, see the sequential combinator in the Hedgehog.Gen module.

executeParallel :: (MonadTest m, MonadCatch m, MonadBaseControl IO m, HasCallStack) => (forall v. state v) -> Parallel m state -> m ()

Executes the prefix actions sequentially, then executes the two branches in parallel, verifying that no exceptions are thrown and that there is at least one sequential interleaving where all the post-conditions are met.

To generate parallel actions to execute, see the parallel combinator in the Hedgehog.Gen module.

data Var a v

Variables are the potential or actual result of executing an action. They are parameterised by either Symbolic or Concrete depending on the phase of the test.

Symbolic variables are the potential results of actions. These are used when generating the sequence of actions to execute. They allow actions which occur later in the sequence to make use of the result of an action which came earlier in the sequence.

Concrete variables are the actual results of actions. These are used during test execution. They provide access to the actual runtime value of a variable.

The state update Callback for a command needs to be polymorphic in the type of variable because it is used in both the generation and the execution phase.

Constructors

Var (v a)

concrete :: Var a Concrete -> a

Take the value from a concrete variable.

opaque :: Var (Opaque a) Concrete -> a

Take the value from an opaque concrete variable.

data Symbolic a

Symbolic values.

newtype Concrete a where

Concrete values.

Constructors

Concrete :: a -> Concrete a

newtype Opaque a

Opaque values.

Useful if you want to put something without a Show instance inside something which you'd like to be able to display.

For example:

  data State v =
    State {
        stateRefs :: [Var (Opaque (IORef Int)) v]
      } deriving (Eq, Show)

Constructors

Opaque

Fields

unOpaque :: a

Transformers

distribute :: (Distributive g, Transformer f g m) => g (f m) a -> f (g m) a

Distribute one monad transformer over another.

Functors

class HTraversable t where

Higher-order traversable functors.

This is used internally to make symbolic variables concrete given an Environment.

Methods

htraverse :: Applicative f => (forall a. g a -> f (h a)) -> t g -> f (t h)

class Eq1 (f :: Type -> Type)

Lifting of the Eq class to unary type constructors.

Since: base-4.9.0.0

Minimal complete definition

liftEq

eq1 :: (Eq1 f, Eq a) => f a -> f a -> Bool

Lift the standard (==) function through the type constructor.

Since: base-4.9.0.0

class Eq1 f => Ord1 (f :: Type -> Type)

Lifting of the Ord class to unary type constructors.

Since: base-4.9.0.0

Minimal complete definition

liftCompare

compare1 :: (Ord1 f, Ord a) => f a -> f a -> Ordering

Lift the standard compare function through the type constructor.

Since: base-4.9.0.0

class Show1 (f :: Type -> Type)

Lifting of the Show class to unary type constructors.

Since: base-4.9.0.0

Minimal complete definition

liftShowsPrec

showsPrec1 :: (Show1 f, Show a) => Int -> f a -> ShowS

Lift the standard showsPrec and showList functions through the type constructor.

Since: base-4.9.0.0