Hedgehog.Internal.Property

Property

data Property

A property test, along with some configurable limits like how many times to run the test.

Constructors

Property

Fields

propertyConfig :: !PropertyConfig

propertyTest :: PropertyT IO ()

newtype PropertyT m a

The property monad transformer allows both the generation of test inputs and the assertion of expectations.

Constructors

PropertyT

Fields

unPropertyT :: TestT (GenT m) a

newtype PropertyName

The name of a property.

Can be constructed using OverloadedStrings:

  "apples" :: PropertyName

Constructors

PropertyName

Fields

unPropertyName :: String

data PropertyConfig

Configuration for a property test.

Constructors

PropertyConfig

Fields

propertyTestLimit :: !TestLimit

propertyDiscardLimit :: !DiscardLimit

propertyShrinkLimit :: !ShrinkLimit

propertyShrinkRetries :: !ShrinkRetries

newtype TestLimit

The number of successful tests that need to be run before a property test is considered successful.

Can be constructed using numeric literals:

  200 :: TestLimit

Constructors

TestLimit Int

newtype DiscardLimit

The number of discards to allow before giving up.

Can be constructed using numeric literals:

  10000 :: DiscardLimit

Constructors

DiscardLimit Int

newtype ShrinkLimit

The number of shrinks to try before giving up on shrinking.

Can be constructed using numeric literals:

  1000 :: ShrinkLimit

Constructors

ShrinkLimit Int

newtype ShrinkRetries

The number of times to re-run a test during shrinking. This is useful if you are testing something which fails non-deterministically and you want to increase the change of getting a good shrink.

If you are doing parallel state machine testing, you should probably set shrink retries to something like 10. This will mean that during shrinking, a parallel test case requires 10 successful runs before it is passes and we try a different shrink.

Can be constructed using numeric literals:

  0 :: ShrinkRetries

Constructors

ShrinkRetries Int

withTests :: TestLimit -> Property -> Property

Set the number times a property should be executed before it is considered successful.

withDiscards :: DiscardLimit -> Property -> Property

Set the number times a property is allowed to discard before the test runner gives up.

withShrinks :: ShrinkLimit -> Property -> Property

Set the number times a property is allowed to shrink before the test runner gives up and prints the counterexample.

withRetries :: ShrinkRetries -> Property -> Property

Set the number times a property will be executed for each shrink before the test runner gives up and tries a different shrink. See ShrinkRetries for more information.

property :: HasCallStack => PropertyT IO () -> Property

Creates a property with the default configuration.

test :: Monad m => TestT m a -> PropertyT m a

Lift a test in to a property.

Because both TestT and PropertyT have MonadTest instances, this function is not often required. It can however be useful for writing functions directly in TestT and thus gaining a MonadTransControl instance at the expense of not being able to generate additional inputs using forAll.

One use case for this is writing tests which use ResourceT:

  property $ do
    n <- forAll $ Gen.int64 Range.linearBounded
    test . runResourceT $ do
      -- test with resource usage here

forAll :: (Monad m, Show a, HasCallStack) => Gen a -> PropertyT m a

Generates a random input for the test by running the provided generator.

forAllT :: (Monad m, Show a, HasCallStack) => GenT m a -> PropertyT m a

Generates a random input for the test by running the provided generator.

forAllWith :: (Monad m, HasCallStack) => (a -> String) -> Gen a -> PropertyT m a

Generates a random input for the test by running the provided generator.

This is a the same as forAll but allows the user to provide a custom rendering function. This is useful for values which don't have a Show instance.

forAllWithT :: (Monad m, HasCallStack) => (a -> String) -> GenT m a -> PropertyT m a

Generates a random input for the test by running the provided generator.

This is a the same as forAllT but allows the user to provide a custom rendering function. This is useful for values which don't have a Show instance.

discard :: Monad m => PropertyT m a

Discards the current test entirely.

Group

data Group

A named collection of property tests.

Constructors

Group

Fields

groupName :: !GroupName

groupProperties :: ![(PropertyName, Property)]

newtype GroupName

The name of a group of properties.

Can be constructed using OverloadedStrings:

  "fruit" :: GroupName

Constructors

GroupName

Fields

unGroupName :: String

TestT

class Monad m => MonadTest m where

Minimal complete definition

liftTest

Methods

liftTest :: Test a -> m a

type Test = TestT Identity

A test monad allows the assertion of expectations.

newtype TestT m a

A test monad transformer allows the assertion of expectations.

Constructors

TestT

Fields

unTest :: ExceptT Failure (WriterT [Log] m) a

data Log

Log messages which are recorded during a test run.

Constructors

Annotation (Maybe Span) String

Footnote String

data Failure

Details on where and why a test failed.

Constructors

Failure (Maybe Span) String (Maybe Diff)

data Diff

The difference between some expected and actual value.

Constructors

Diff

Fields

diffPrefix :: String

diffRemoved :: String

diffInfix :: String

diffAdded :: String

diffSuffix :: String

diffValue :: ValueDiff

annotate :: (MonadTest m, HasCallStack) => String -> m ()

Annotates the source code with a message that might be useful for debugging a test failure.

annotateShow :: (MonadTest m, Show a, HasCallStack) => a -> m ()

Annotates the source code with a value that might be useful for debugging a test failure.

footnote :: MonadTest m => String -> m ()

Logs a message to be displayed as additional information in the footer of the failure report.

footnoteShow :: (MonadTest m, Show a) => a -> m ()

Logs a value to be displayed as additional information in the footer of the failure report.

failure :: (MonadTest m, HasCallStack) => m a

Causes a test to fail.

success :: MonadTest m => m ()

Another name for pure ().

assert :: (MonadTest m, HasCallStack) => Bool -> m ()

Fails the test if the condition provided is False.

(===) :: (MonadTest m, Eq a, Show a, HasCallStack) => a -> a -> m () infix 4

Fails the test if the two arguments provided are not equal.

eval :: (MonadTest m, HasCallStack) => a -> m a

Fails the test if the value throws an exception when evaluated to weak head normal form (WHNF).

evalM :: (MonadTest m, MonadCatch m, HasCallStack) => m a -> m a

Fails the test if the action throws an exception.

The benefit of using this over simply letting the exception bubble up is that the location of the closest evalM will be shown in the output.

evalIO :: (MonadTest m, MonadIO m, HasCallStack) => IO a -> m a

Fails the test if the IO action throws an exception.

The benefit of using this over liftIO is that the location of the exception will be shown in the output.

evalEither :: (MonadTest m, Show x, HasCallStack) => Either x a -> m a

Fails the test if the Either is Left, otherwise returns the value in the Right.

evalExceptT :: (MonadTest m, Show x, HasCallStack) => ExceptT x m a -> m a

Fails the test if the ExceptT is Left, otherwise returns the value in the Right.

Internal

These functions are exported in case you need them in a pinch, but are not part of the public API and may change at any time, even as part of a minor update.

defaultConfig :: PropertyConfig

The default configuration for a property test.

mapConfig :: (PropertyConfig -> PropertyConfig) -> Property -> Property

Map a config modification function over a property.

failDiff :: (MonadTest m, Show a, Show b, HasCallStack) => a -> b -> m ()

Fails with an error which shows the difference between two values.

failException :: (MonadTest m, HasCallStack) => SomeException -> m a

Fails with an error which renders the type of an exception and its error message.

failWith :: (MonadTest m, HasCallStack) => Maybe Diff -> String -> m a

Fail the test with an error message, useful for building other failure combinators.

writeLog :: MonadTest m => Log -> m ()

Log some information which might be relevant to a potential test failure.

mkTest :: (Either Failure a, [Log]) -> Test a

mkTestT :: m (Either Failure a, [Log]) -> TestT m a

runTest :: Test a -> (Either Failure a, [Log])

runTestT :: TestT m a -> m (Either Failure a, [Log])