Hedgehog.Internal.Report

Report

data Summary

A summary of all the properties executed.

Constructors

Summary

Fields

summaryWaiting :: !PropertyCount

summaryRunning :: !PropertyCount

summaryFailed :: !PropertyCount

summaryGaveUp :: !PropertyCount

summaryOK :: !PropertyCount

data Report a

A report on a running or completed property test.

Constructors

Report

Fields

reportTests :: !TestCount

reportDiscards :: !DiscardCount

reportStatus :: !a

data Progress

The status of a running property test.

Constructors

Running

Shrinking !FailureReport

data Result

The status of a completed property test.

In the case of a failure it provides the seed used for the test, the number of shrinks, and the execution log.

Constructors

Failed !FailureReport

GaveUp

OK

data FailureReport

Constructors

FailureReport

Fields

failureSize :: !Size

failureSeed :: !Seed

failureShrinks :: !ShrinkCount

failureAnnotations :: ![FailedAnnotation]

failureLocation :: !(Maybe Span)

failureMessage :: !String

failureDiff :: !(Maybe Diff)

failureFootnotes :: ![String]

data FailedAnnotation

Constructors

FailedAnnotation

Fields

failedSpan :: !(Maybe Span)

failedValue :: !String

newtype ShrinkCount

The numbers of times a property was able to shrink after a failing test.

Constructors

ShrinkCount Int

newtype TestCount

The number of tests a property ran successfully.

Constructors

TestCount Int

newtype DiscardCount

The number of tests a property had to discard.

Constructors

DiscardCount Int

newtype PropertyCount

The number of properties in a group.

Constructors

PropertyCount Int

data Style

Constructors

StyleDefault

StyleAnnotation

StyleFailure

data Markup

Constructors

WaitingIcon

WaitingHeader

RunningIcon

RunningHeader

ShrinkingIcon

ShrinkingHeader

FailedIcon

FailedHeader

GaveUpIcon

GaveUpHeader

SuccessIcon

SuccessHeader

DeclarationLocation

StyledLineNo !Style

StyledBorder !Style

StyledSource !Style

AnnotationGutter

AnnotationValue

FailureArrows

FailureGutter

FailureMessage

DiffPrefix

DiffInfix

DiffSuffix

DiffSame

DiffRemoved

DiffAdded

ReproduceHeader

ReproduceGutter

ReproduceSource

renderProgress :: MonadIO m => Maybe UseColor -> Maybe PropertyName -> Report Progress -> m String

renderResult :: MonadIO m => Maybe UseColor -> Maybe PropertyName -> Report Result -> m String

renderSummary :: MonadIO m => Maybe UseColor -> Summary -> m String

renderDoc :: MonadIO m => Maybe UseColor -> Doc Markup -> m String

ppProgress :: MonadIO m => Maybe PropertyName -> Report Progress -> m (Doc Markup)

ppResult :: MonadIO m => Maybe PropertyName -> Report Result -> m (Doc Markup)

ppSummary :: MonadIO m => Summary -> m (Doc Markup)

fromResult :: Result -> Summary

Construct a summary from a single result.

mkFailure :: Size -> Seed -> ShrinkCount -> Maybe Span -> String -> Maybe Diff -> [Log] -> FailureReport