Hedgehog.Internal.Tree

Documentation

newtype Tree m a

An effectful tree, each node in the tree can have an effect before it is produced.

Constructors

Tree

Fields

runTree :: m (Node m a)

data Node m a

A node in an effectful tree, as well as its unevaluated children.

Constructors

Node

Fields

nodeValue :: a

nodeChildren :: [Tree m a]

fromNode :: Applicative m => Node m a -> Tree m a

Create a Tree from a Node

unfold :: Monad m => (a -> [a]) -> a -> Tree m a

Create a tree from a value and an unfolding function.

unfoldForest :: Monad m => (a -> [a]) -> a -> [Tree m a]

Create a forest from a value and an unfolding function.

expand :: Monad m => (a -> [a]) -> Tree m a -> Tree m a

Expand a tree using an unfolding function.

prune :: Monad m => Tree m a -> Tree m a

Throw away a tree's children.

render :: Monad m => Tree m String -> m String

Render a tree of strings, note that this forces all the delayed effects in the tree.